requirements.txt
asyncua
psycopg2-binary
